requirements.txt
pywebview
sqlalchemy
pyinstaller
pywin32